Add a flat-file implementation of property_set / property_get.
For non-android systems, implement property_set() and property_get()
in terms of operations on a flat file of key=value pairs. The path is
hard-coded in include.mk to $sysconfdir/adb.properties.

The implementation is not particularly efficient, but designed to be
safe against updates from multiple threads and processes.